Reworked code to support developer implementations of IClientAssertionCertificate with an X509 claim.
Added unit test to check for c5x claim in JWT request header.
