+++ src/Microsoft.IdentityModel.Clients.ActiveDirectory/IClientAssertionCertificate.cs
@@ -48,10 +48,5 @@ namespace Microsoft.IdentityModel.Clients.ActiveDirectory
         /// Thumbprint of the Certificate
         /// </summary>
         string Thumbprint { get; }
-
-        /// <summary>
-        /// Returns the base64 encoded certificate
-        /// </summary>
-        string GetEncodedCertificate();
     }
 }
