+++ tests/Test.ADAL.NET.Common/Mocks/MockHttpMessageHandler.cs
@@ -60,6 +60,8 @@ namespace Test.ADAL.NET.Common.Mocks
             this.Url = Url;
         }
 
+        public object AdditionalValidationOutput;
+
         protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
         {
             Assert.AreEqual(Method, request.Method);
